[세 번째 요구사항 추가하기 - 책 통계] 35강. 애플리케이션 대신 DB로 기능 구현하기

+++ src/main/kotlin/com/group/libraryapp/controller/book/BookController.kt
@@ -29,7 +29,7 @@ class BookController(private val bookService: BookService) {
     }
 
     @GetMapping("/book/loan")
-    fun countLoanedBook(): Int {
+    fun countLoanedBook(): Long {
         return bookService.countLoanedBook()
     }
